+++ src/apps/Elsa.Server.Web/appsettings.json
@@ -25,7 +25,7 @@
   },
   "DatabaseProvider": "Sqlite",
   "LoggingFramework": {
-    "Defaults": ["Console"],
+    "Defaults": ["ConsoleInfo", "ConsoleDebugColored"],
     "Sinks": [
       {
         "Type": "Console",
