+++ src/modules/Elsa.Logging.Core/Extensions/LoggingBuilderExtensions.cs
@@ -1,12 +1,13 @@
+using Elsa.Logging.Models;
 using Microsoft.Extensions.Logging;
 
 namespace Elsa.Logging.Extensions;
 
 public static class LoggingBuilderExtensions
 {
-    public static ILoggingBuilder AddCategoryFilters(this ILoggingBuilder builder, SinkOptions.SinkOptions sinkOptions)
+    public static ILoggingBuilder AddCategoryFilters(this ILoggingBuilder builder, LogSinkOptions logSinkOptions)
     {
-        return builder.AddCategoryFilters(sinkOptions.CategoryFilters);
+        return builder.AddCategoryFilters(logSinkOptions.CategoryFilters);
     }
 
     public static ILoggingBuilder AddCategoryFilters(this ILoggingBuilder builder, IDictionary<string, LogLevel>? categoryFilters)
